Feat: creator info 페이지 context, reduce 추가 완료
