+++ src/router/router.js
@@ -5,6 +5,7 @@ import SignUp from '../pages/SignUp';
 import UserPage from '../pages/User';
 import Searched from '../pages/Searched';
 import UserSetting from '../pages/UserSetting';
+import { Provider } from '../context/creator/creatorContext';
 import App from '../App';
 
 const Path = () => {
@@ -14,9 +15,11 @@ const Path = () => {
       <Route path="/" exact component={App} />
       <Route path="/signin" render={() => (hasToken ? <App /> : <SignIn />)} />
       <Route path="/signup" render={() => (hasToken ? <App /> : <SignUp />)} />
-      <Route path="/user" component={UserPage} />
       <Route path="/searched" component={Searched} />
-      <Route path="/usersetting" component={UserSetting} />
+      <Provider>
+        <Route path="/user" component={UserPage} />
+        <Route path="/usersetting" component={UserSetting} />
+      </Provider>
     </Router>
   );
 };
